chore: remove uncessary code
